Sort event data by location on discover fragment

+++ app/src/main/java/net/slc/hoga/bantuin/Model/Event.java
@@ -7,8 +7,8 @@ public class Event {
     private String key;
     private String title;
     private String description;
-    private Long lat;
-    private Long lng;
+    private Double lat;
+    private Double lng;
     private String location;
     private User organizer;
     private String date;
@@ -45,19 +45,19 @@ public class Event {
         this.description = description;
     }
 
-    public Long getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Long lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public Long getLng() {
+    public Double getLng() {
         return lng;
     }
 
-    public void setLng(Long lng) {
+    public void setLng(Double lng) {
         this.lng = lng;
     }
